Implement feedback from @Widcket
