+++ internal/cli/importcmd/apis.go
@@ -1,3 +1,3 @@
-package cli
+package importcmd
 
 // Put here the APIs handler logic
