+++ internal/cli/importcmd/apps.go
@@ -1,3 +1,3 @@
-package cli
+package importcmd
 
 // Put here the Apps (Clients) handler logic
