+++ internal/cli/importcmd/roles.go
@@ -1,3 +1,3 @@
-package cli
+package importcmd
 
 // Put here the Roles handler logic
